Dropped Landscape Support
Originally, I wanted to have ScrollControl rotate for landscape,
however I found two problems with this.

1.) While using a ScrollControl to navigate a page in portrait mode is
awesome, it is a bit too small to be useful in most cases in landscape.

2.) - (void)didRotateFromInterfaceOrientation:(int)fp8 is the function
that gets called when the device “rotates”, however, sometimes the
UIScrollView is rotated before the function is called and sometimes
after, making it very difficult to dynamically set the frame for
ScrollControl based off the UIScrollView.

So, instead I simply disabled ScrollControl in landscape, and reenabled
once the user reenters portrait mode.